+++ src/ios/Diagnostic.h
@@ -22,6 +22,7 @@ extern NSString*const AUTHORIZATION_DENIED;
 extern NSString*const AUTHORIZATION_GRANTED;
 extern NSString*const AUTHORIZATION_PROVISIONAL;
 extern NSString*const AUTHORIZATION_EPHEMERAL;
+extern NSString*const AUTHORIZATION_LIMITED;
 
 @interface Diagnostic : CDVPlugin
 
@@ -33,10 +34,16 @@ extern NSString*const AUTHORIZATION_EPHEMERAL;
 - (void) switchToSettings: (CDVInvokedUrlCommand*)command;
 - (void) getBackgroundRefreshStatus: (CDVInvokedUrlCommand*)command;
 - (void) getArchitecture: (CDVInvokedUrlCommand*)command;
+- (void) getCurrentBatteryLevel: (CDVInvokedUrlCommand*)command;
+- (void) getDeviceOSVersion: (CDVInvokedUrlCommand*)command;
+- (void) getBuildOSVersion: (CDVInvokedUrlCommand*)command;
+- (void) isMobileDataEnabled: (CDVInvokedUrlCommand*)command;
 
 // Utilities
 + (id) getInstance;
 - (void) sendPluginResult: (CDVPluginResult*)result :(CDVInvokedUrlCommand*)command;
+- (void) sendPluginResultSuccess:(CDVInvokedUrlCommand*)command;
+- (void) sendPluginNoResultAndKeepCallback:(CDVInvokedUrlCommand*)command;
 - (void) sendPluginResultBool: (BOOL)result :(CDVInvokedUrlCommand*)command;
 - (void) sendPluginResultString: (NSString*)result :(CDVInvokedUrlCommand*)command;
 - (void) sendPluginError: (NSString*) errorMessage :(CDVInvokedUrlCommand*)command;