+++ src/ios/Diagnostic_Bluetooth.m
@@ -65,20 +65,17 @@ static NSString*const LOG_TAG = @"Diagnostic_Bluetooth[native]";
 {
     [self.commandDelegate runInBackground:^{
         @try {
-            NSString* state = [self getBluetoothState];
-
-            if([state isEqual: @"unauthorized"]){
-                /*
-                 When the application requests to start scanning for bluetooth devices that is when the user is presented with a consent dialog.
-                 */
-                [diagnostic logDebug:@"Requesting bluetooth authorization"];
-                [self ensureBluetoothManager];
-                [self.bluetoothManager scanForPeripheralsWithServices:nil options:nil];
-                [self.bluetoothManager stopScan];
-            }else{
-                [diagnostic logDebug:@"Bluetooth authorization is already granted"];
+            
+            NSString* authState = [self getAuthorizationStatus];
+            
+            if([authState isEqual:AUTHORIZATION_GRANTED]){
+                [diagnostic sendPluginError:@"Bluetooth authorization is already granted" :command];
+            }else if([authState isEqual:AUTHORIZATION_DENIED]){
+                [diagnostic sendPluginError:@"Bluetooth authorization has been denied" :command];
+            }else{ // AUTHORIZATION_NOT_DETERMINED
+                [self ensureBluetoothManager]; // invoke Bluetooth manager to trigger permission dialog
+                [diagnostic sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK] :command];
             }
-            [diagnostic sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK] :command];
         }
         @catch (NSException *exception) {
             [diagnostic handlePluginException:exception :command];
@@ -98,6 +95,17 @@ static NSString*const LOG_TAG = @"Diagnostic_Bluetooth[native]";
         }
     }];
 
+}
+
+- (void) getAuthorizationStatus: (CDVInvokedUrlCommand*)command {
+    [self.commandDelegate runInBackground:^{
+        @try {
+            NSString* authState = [self getAuthorizationStatus];
+            [diagnostic sendPluginResultString:authState :command];
+        }@catch (NSException *exception) {
+            [diagnostic handlePluginException:exception :command];
+        }
+    }];
 }
 
 /********************************/
@@ -155,6 +163,40 @@ static NSString*const LOG_TAG = @"Diagnostic_Bluetooth[native]";
                                  options:@{CBCentralManagerOptionShowPowerAlertKey: @(NO)}];
         [self centralManagerDidUpdateState:self.bluetoothManager]; // Send initial state
     }
+}
+
+// https://stackoverflow.com/q/57238647/777265
+- (NSString*) getAuthorizationStatus {
+    NSString* authState;
+    if (@available(iOS 13.0, *)){
+        CBManagerAuthorization authorization;
+        if(@available(iOS 13.1, *)){
+            authorization = CBCentralManager.authorization;
+        }else{
+            // iOS 13.0 requires BT manager to be initialized in order to check authorization which results in BT permissions dialog if not already authorized
+            [self ensureBluetoothManager];
+            authorization = [self.bluetoothManager authorization];
+        }
+        
+        switch(authorization){
+            case CBManagerAuthorizationAllowedAlways:
+                authState = AUTHORIZATION_GRANTED;
+                break;
+            case CBManagerAuthorizationDenied:
+                authState = AUTHORIZATION_DENIED;
+                break;
+            case CBManagerAuthorizationRestricted:
+                authState = AUTHORIZATION_DENIED;
+                break;
+            case CBManagerAuthorizationNotDetermined:
+                authState = AUTHORIZATION_NOT_DETERMINED;
+                break;
+        }
+    }else{
+        // Device is running iOS <13.0 so doesn't require Bluetooth permission at run-time
+        authState = AUTHORIZATION_GRANTED;
+    }
+    return authState;
 }
 
 /********************************/